Re-formatting and clean up. Still have some errors regarding the scenes
