+++ app/scripts/people/container.js
