+++ app/scripts/people/dialogue.js
