+++ app/scripts/people/logo.js
