+++ app/scripts/people/person.js
