+++ app/scripts/structure/control-button.js
