+++ app/scripts/structure/display.js
